Fix bug
The bug is for Google Map InfoBox do deep comparison for selectedItem but the actual item got re-rendered because the change of state. So what the old code do is it compared 2 different objects address and see if they are similar (and they aren't, the data structure is the same but it is the copied of the old one). The new code use their `id` field instead (if they are the same, eslint will yelled at me :-P)